requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
Werkzeug==2.3.7
SQLAlchemy==2.0.23
gunicorn==21.2.0
